added ScheduleCallModal, and other UI fixes
