+++ package.json
@@ -53,10 +53,11 @@
     "framer-motion": "^11.0.0",
     "input-otp": "^1.2.4",
     "lucide-react": "^0.446.0",
-    "next": "13.5.1",
+    "next": "^15.4.5",
     "next-themes": "^0.3.0",
-    "postcss": "8.4.30",
+    "postcss": "^8.5.6",
     "react": "18.2.0",
+    "react-calendly": "^4.4.0",
     "react-day-picker": "^8.10.1",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.53.0",
